Support different commands
In an attemt to stay within the boundaries of this submodule, this hack-ish
solution creates multiple configs and requests depending on the "command"
specified. This is because bulbs, as-is, won't support a base-URI that doesn't
include the DB name. Since OrientDB requires the command BEFORE the DB name,
this is necessary (unless bulbs request gets re-written, or we don't use it
entirely).